+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/BindingContainerTransformerTest.kt
@@ -617,7 +617,10 @@ class BindingContainerTransformerTest : MetroCompilerTest() {
     ) {
       ExampleGraph.generatedImpl().createGraphWithNoArgs()
       assertDiagnostics(
-        "w: Included Dagger module 'ExampleModule' declares a `subcomponents` parameter but this will be ignored by Metro in interop."
+        """
+        w: Included Dagger module 'ExampleModule' declares a `subcomponents` parameter but this will be ignored by Metro in interop.
+        """
+          .trimIndent()
       )
     }
   }